Change spell visual kit fields to signed.

+++ WowPacketParserModule.V6_0_2_19033/Parsers/SpellHandler.cs
@@ -535,8 +535,8 @@ namespace WowPacketParserModule.V6_0_2_19033.Parsers
         {
             PlaySpellVisualKit visualKitData = new PlaySpellVisualKit();
             visualKitData.Guid = packet.ReadPackedGuid128("Unit");
-            visualKitData.KitId = (uint)packet.ReadInt32("KitRecID");
-            visualKitData.KitType = (uint)packet.ReadInt32("KitType");
+            visualKitData.KitId = packet.ReadInt32("KitRecID");
+            visualKitData.KitType = packet.ReadInt32("KitType");
             visualKitData.Duration = packet.ReadUInt32("Duration");
             visualKitData.Time = packet.Time;
             Storage.SpellPlayVisualKit.Add(visualKitData);
